tercer entregable

+++ src/main/java/com/nttdata/bootcamp/msclient/infrastructure/ClientRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientRepository extends ReactiveMongoRepository<Client, Integer> {
+public interface ClientRepository extends ReactiveMongoRepository<Client, Long> {
 }
